updates package definition to match standards
update version to 1.0.2 because maven seems stuck
