+++ pom.xml
@@ -11,7 +11,7 @@
   <groupId>testrail</groupId>
   <artifactId>testrail</artifactId>
   <name>TestRail Plugin</name>
-  <version>1.0.1-SNAPSHOT</version>
+  <version>1.0.2-SNAPSHOT</version>
   <licenses>
     <license>
       <name>Apache License, Version 2.0</name>
