+++ src/main/java/org/jenkinsci/plugins/testrail/JunitResults/JUnitResults.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package testrail.testrail.JunitResults;
+package org.jenkinsci.plugins.testrail.JunitResults;
 
 import hudson.FilePath;
 import hudson.remoting.VirtualChannel;
