+++ src/main/java/org/jenkinsci/plugins/testrail/TestRailObjects/Case.java
@@ -16,13 +16,43 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package testrail.testrail.TestRailObjects;
+package org.jenkinsci.plugins.testrail.TestRailObjects;
+
 
 /**
  * Created by Drew on 3/24/2014.
  */
-public class ElementNotFoundException extends Exception {
-    public ElementNotFoundException(String message) {
-        super(message);
+public class Case {
+    private int id;
+    private String title;
+    private int sectionId;
+    private String refs;
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public void setTitle(String title) {
+        this.title = title.trim();
+    }
+
+    public void setSectionId(int sectionId) { this.sectionId = sectionId; }
+
+    public void setRefs(String refs) {
+        this.refs = refs;
+    }
+
+    public int getId() {
+        return this.id;
+    }
+
+    public String getTitle() {
+        return this.title;
+    }
+
+    public int getSectionId() { return this.sectionId; }
+
+    public String getRefs() {
+        return this.refs;
     }
 }
