+++ src/main/java/org/jenkinsci/plugins/testrail/TestRailObjects/Suite.java
@@ -16,31 +16,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package testrail.testrail.JunitResults;
-
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
+package org.jenkinsci.plugins.testrail.TestRailObjects;
 
 /**
  * Created by Drew on 3/24/2014.
  */
-public class Testcase {
+public class Suite {
+    private int id;
     private String name;
-    private Failure failure;
-    private Float time;
-    private String refs;
 
-    @XmlAttribute
-    public void setName(String name) { this.name = name.trim(); }
-    @XmlElement(name = "failure")
-    public void setFailure(Failure failure) { this.failure = failure; }
-    @XmlAttribute(name = "time")
-    public void setTime(Float time) { this.time = time; }
-    @XmlAttribute(name = "refs")
-    public void setRefs(String refs) { this.refs = refs; }
+    public void setId(int id) {
+        this.id = id;
+    }
 
-    public String getName() { return this.name; }
-    public Failure getFailure() { return this.failure; }
-    public Float getTime() { return this.time; }
-    public String getRefs() { return this.refs; }
+    public void setName(String name) {
+        this.name = name.trim();
+    }
+
+    public int getId() {
+        return this.id;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    public String getStringId() { return Integer.toString(id); }
+
 }
