+++ src/main/java/org/jenkinsci/plugins/testrail/TestRailResponse.java
@@ -16,27 +16,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package testrail.testrail.JunitResults;
-
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlValue;
+package org.jenkinsci.plugins.testrail;
 
 /**
- * Created by Drew on 3/26/2014.
+ * Created by Drew on 3/20/14.
  */
-public class Failure {
-    private String type;
-    private String message;
-    private String text;
+public class TestRailResponse {
+    private Integer status;
+    public Integer getStatus() { return status; }
+    
+    private String body;
+    public String getBody() { return body; }
 
-    @XmlAttribute
-    public void setType(String type) { this.type = type; }
-    @XmlAttribute
-    public void  setMessage(String message) { this.message = message; }
-    @XmlValue
-    public void setText(String text) { this.text = text; }
-
-    public String getType() { return this.type; }
-    public String getMessage() { return this.message; }
-    public String getText() { return this.text; }
+    public TestRailResponse(Integer status, String body) {
+        this.status = status;
+        this.body = body;
+    }
 }
