+++ src/main/resources/org/jenkinsci/plugins/testrail/TestRailNotifier/global.jelly
